refact:Refactoring scaffold project
