+++ src/app.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
-import { ProductsModule } from './products/products.module';
-import { AuthModule } from './auth/auth.module';
-import { CategoriesModule } from './categories/categories.module';
-import { PaymentsModule } from './payments/payments.module';
-import { OrdersModule } from './orders/orders.module';
-import { CartModule } from './cart/cart.module';
-import { UsersModule } from './users/users.module';
+import { ProductsModule } from './modules/products/products.module';
+import { AuthModule } from './modules/auth/auth.module';
+import { CategoriesModule } from './modules/categories/categories.module';
+import { PaymentsModule } from './modules/payments/payments.module';
+import { OrdersModule } from './modules/orders/orders.module';
+import { CartModule } from './modules/cart/cart.module';
+import { UsersModule } from './modules/users/users.module';
 import { ThrottlerModule } from '@nestjs/throttler';
-import { SubcategoryModule } from './subcategory/subcategory.module';
+import { SubcategoryModule } from './modules/subcategory/subcategory.module';
 import { GraphQLModule } from '@nestjs/graphql';
 import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo';
 import { join } from 'path';
