--- src/auth/dto/create-auth.dto.ts
@@ -1,3 +1,0 @@
-import { CreateUserDto } from 'src/users/dto/create-user.dto';
-
-export class CreateAuthDto extends CreateUserDto {}
