+++ src/config/swagger.config.ts
@@ -1,0 +1,8 @@
+import { DocumentBuilder } from "@nestjs/swagger";
+
+export const config = new DocumentBuilder()
+  .setTitle('Ecommerce NestJS')
+  .setDescription('Ecommerce API NESTJS')
+  .setVersion('1.0')
+  .addBearerAuth()
+  .build();
