+++ src/main.ts
@@ -1,7 +1,8 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { ValidationPipe } from '@nestjs/common';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import { SwaggerModule } from '@nestjs/swagger';
+import { config } from './config/swagger.config';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
@@ -18,12 +19,6 @@ async function bootstrap() {
     origin: process.env.ORIGIN_CLIENT,
   });
 
-  const config = new DocumentBuilder()
-    .setTitle('Ecommerce NestJS')
-    .setDescription('Ecommerce API NESTJS')
-    .setVersion('1.0')
-    .addBearerAuth()
-    .build();
   const documentFactory = () => SwaggerModule.createDocument(app, config);
   SwaggerModule.setup('/api/v1/docs', app, documentFactory);
   await app.listen(process.env.PORT ?? 4001);
