+++ src/modules/auth/auth.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Post, Body } from '@nestjs/common';
 import { AuthService } from './auth.service';
-import { CreateUserDto } from 'src/users/dto/create-user.dto';
+import { CreateUserDto } from 'src/modules/users/dto/create-user.dto';
 import { LoginAuthDto } from './dto/login-auth.dto';
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
