+++ src/modules/auth/dto/login-auth.dto.ts
@@ -2,7 +2,7 @@ import { OmitType } from '@nestjs/mapped-types';
 import { ApiProperty } from '@nestjs/swagger';
 import { Transform } from 'class-transformer';
 import { IsEmail, IsNotEmpty, IsString, MinLength } from 'class-validator';
-import { CreateUserDto } from 'src/users/dto/create-user.dto';
+import { CreateUserDto } from 'src/modules/users/dto/create-user.dto';
 
 export class LoginAuthDto extends OmitType(CreateUserDto, [
   'address',
