+++ src/modules/categories/categories.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { CategoriesService } from './categories.service';
-import { PrismaModule } from 'src/prisma/prisma.module';
+import { PrismaModule } from 'src/common/prisma/prisma.module';
 import { CategoriesResolver } from './categories.resolver';
 
 @Module({
