+++ src/modules/subcategory/entities/subcategory.entity.ts
